Correctly up(?)grade to Anaconda from Miniconda.
Note that (Mini)conda and Anaconda versions are not synced in any
straightforward way: by trial and error, we learn that Anaconda 2.5
corresponds to conda 3.19.0.

Note also that switching to Anaconda switches to using MKL, even if nomkl was
previous installed, so we need to go to some effort to revert that.